Annual Decreae in TB from Single Estimate for Each Group

From a vetor of assumed annual decreases in TB, this function constructs a matrix with yearly estimates of TB decrease

TB_decrease_as_matrix(TB_decrease, TB_cohorts = c("15-19 years",
  "20-24 years", "25-29 years"), data_start = 1969,
  projected_cohort = 9, cohort_length = 5)

Arguments

TB_decrease
A named vector of percentage decreases (i.e 1%) in TB incidence. Names should refer to age group.

TB_cohorts
A character vector indicating the names of the cohorts modelled. Defaults to 15-19 years, 20-24 years, and 25-29 years.

data_start
Numeric, the year the first data point originates, defaults to 1969 as in Sutherland et al.

projected_cohort
Numeric, the number of cohort for which to simulate, defaults to 5. Combines with cohort_length to give the length of time to simulate.

cohort_length
Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.

Value

A matrix of annual TB incidence decreases as a percentage (0.01) for a given number of years and age groups

Examples

TB_decrease <- c(10, 8, 11)

TB_decrease_as_matrix(TB_decrease)

#>      [,1] [,2] [,3]
#> 1969  0.1 0.08 0.11
#> 1970  0.1 0.08 0.11
#> 1971  0.1 0.08 0.11
#> 1972  0.1 0.08 0.11
#> 1973  0.1 0.08 0.11
#> 1974  0.1 0.08 0.11
#> 1975  0.1 0.08 0.11
#> 1976  0.1 0.08 0.11
#> 1977  0.1 0.08 0.11
#> 1978  0.1 0.08 0.11
#> 1979  0.1 0.08 0.11
#> 1980  0.1 0.08 0.11
#> 1981  0.1 0.08 0.11
#> 1982  0.1 0.08 0.11
#> 1983  0.1 0.08 0.11
#> 1984  0.1 0.08 0.11
#> 1985  0.1 0.08 0.11
#> 1986  0.1 0.08 0.11
#> 1987  0.1 0.08 0.11
#> 1988  0.1 0.08 0.11
#> 1989  0.1 0.08 0.11
#> 1990  0.1 0.08 0.11
#> 1991  0.1 0.08 0.11
#> 1992  0.1 0.08 0.11
#> 1993  0.1 0.08 0.11
#> 1994  0.1 0.08 0.11
#> 1995  0.1 0.08 0.11
#> 1996  0.1 0.08 0.11
#> 1997  0.1 0.08 0.11
#> 1998  0.1 0.08 0.11
#> 1999  0.1 0.08 0.11
#> 2000  0.1 0.08 0.11
#> 2001  0.1 0.08 0.11
#> 2002  0.1 0.08 0.11
#> 2003  0.1 0.08 0.11
#> 2004  0.1 0.08 0.11
#> 2005  0.1 0.08 0.11
#> 2006  0.1 0.08 0.11
#> 2007  0.1 0.08 0.11
#> 2008  0.1 0.08 0.11
#> 2009  0.1 0.08 0.11
#> 2010  0.1 0.08 0.11
#> 2011  0.1 0.08 0.11
#> 2012  0.1 0.08 0.11
#> 2013  0.1 0.08 0.11