Function to Calculate Summary Measures from Multiple Runs of the Sutherland et al.'s Model

Function to Calculate Summary Measures from Multiple Runs of the Sutherland et al.'s Model

calc_model_sum_stat(x, ci_level = 0.95, ci_space = "-",
  round_level = 0)

Arguments

x
A list of multiple simulation runs of a singl table as produced by AssessBCGPolicyChange.

ci_level
Numeric, specifies the confidence level to use. Defaults to 0.95.

ci_space
Character, string to use to seperate confidence intervals, defaults to "-"

round_level
Integer, the number of digits to which to round the summarised results. Defaults to 0

Value

A list of tables for a set of summary measures. These are; mean with CI, median, 2.5 97.5

Examples

## Run multiple model simulations
model_sims <- simulate_sutherland_model(cores = 1)

#> Warning: executing %dopar% sequentially: no parallel backend registered

## Summarise the Total effects of ending the scheme table
calc_model_sum_stat(model_sims[["Table 5 - Total Effects of ending the schools BCG scheme"]])

#> $main
#>      1983   1988  1993  1998  2003  2008  2013
#> 1971 "1059" "955" "621" "348" "198" "123" "78"
#> 1976 "688"  "696" "563" "348" "198" "123" "78"
#> 1981 "463"  "420" "379" "306" "198" "123" "78"
#> 1986 "463"  "296" "226" "205" "175" "123" "78"
#> 1991 "463"  "296" "166" "131" "126" "111" "78"
#> 1996 "463"  "296" "166" "91"  "76"  "80"  "70"
#> 2001 "463"  "296" "166" "91"  "52"  "48"  "51"
#> 2006 "463"  "296" "166" "91"  "52"  "33"  "32"
#> 2011 "463"  "296" "166" "91"  "52"  "33"  "22"
#> 
#> $median
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`2.5`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`97.5`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`25`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`75`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $mean
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $lower
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $upper
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#>