Function to Generate Annual Change Scenarios Based on Sutherland Assumptions

Function to Generate Annual Change Scenarios Based on Sutherland Assumptions

gen_annual_change_scenarios(TB_decrease = c(0.09, 0.05, 0.01),
  cohorts_to_sim = 11, samples = 10)

Arguments

TB_decrease
A numeric vector of percentage decreases in TB notifications rates (formated as 0.09 etc.).

cohorts_to_sim
The number of cohorts to simulate. Defaults to 11

samples
Numeric, the number of samples to generate for each scenario. Defaults to 10

Value

A dataframe with each row representing a sample, with annual decrease stored as a matrix

See also

Examples

gen_annual_change_scenarios()

#> # A tibble: 30 x 3
#>    scenario    sample annual_TB_decrease
#>    <chr>        <int> <list>            
#>  1 9% decrease      1 <dbl [55 × 3]>    
#>  2 9% decrease      2 <dbl [55 × 3]>    
#>  3 9% decrease      3 <dbl [55 × 3]>    
#>  4 9% decrease      4 <dbl [55 × 3]>    
#>  5 9% decrease      5 <dbl [55 × 3]>    
#>  6 9% decrease      6 <dbl [55 × 3]>    
#>  7 9% decrease      7 <dbl [55 × 3]>    
#>  8 9% decrease      8 <dbl [55 × 3]>    
#>  9 9% decrease      9 <dbl [55 × 3]>    
#> 10 9% decrease     10 <dbl [55 × 3]>    
#> # … with 20 more rows