Function to sample Sutherland Incidence rates to Include Uncertainty

Function to sample Sutherland Incidence rates to Include Uncertainty

sample_incidence_rates(incidence_rates, population_data, samples = 100)

Arguments

incidence_rates
Incidence rates as formated by sutherland_incidence_rates

population_data
Population data as formated by sutherland_data

samples
Numeric, the number of samples to take. Defaults to 100.

Value

A list of lists. Each list contains matrices of sampled incidence rates for unvaccinated, vaccinated, tuberculin postive and inelgible populations.

Examples

sample_incidence_rates(sutherland_incidence_rates, sutherland_data, samples = 5)

#> [[1]]
#> [[1]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     24.221845    20.14714    10.18676
#> 1972-1976   12.137628    10.70028     0.00000
#> 1977-1981    6.574622     0.00000     0.00000
#> 
#> [[1]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.533937    4.751131     3.21267
#> 1972-1976    2.747253    2.786499     0.00000
#> 1977-1981    1.590595    0.000000     0.00000
#> 
#> [[1]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     21.687908   15.396011    6.974088
#> 1972-1976    9.390375    7.913778    0.000000
#> 1977-1981    4.984027    0.000000    0.000000
#> 
#> [[1]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     15.350877    9.398496    3.759398
#> 1972-1976   10.932427    2.186485    0.000000
#> 1977-1981    6.206334    0.000000    0.000000
#> 
#> [[1]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     24.221845    20.14714    10.18676
#> 1972-1976   12.137628    10.70028     0.00000
#> 1977-1981    6.574622     0.00000     0.00000
#> 
#> 
#> [[2]]
#> [[2]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     26.259196    19.01528    12.45048
#> 1972-1976   10.700277    11.81822     0.00000
#> 1977-1981    7.853021     0.00000     0.00000
#> 
#> [[2]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.624434    4.253394    4.117647
#> 1972-1976    2.511774    2.786499    0.000000
#> 1977-1981    1.694329    0.000000    0.000000
#> 
#> [[2]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.634762   14.761886    8.332834
#> 1972-1976    8.188503    9.031717    0.000000
#> 1977-1981    6.158691    0.000000    0.000000
#> 
#> [[2]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.676692    7.832080    4.699248
#> 1972-1976   13.118912    4.737385    0.000000
#> 1977-1981    7.536263    0.000000    0.000000
#> 
#> [[2]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     26.259196    19.01528    12.45048
#> 1972-1976   10.700277    11.81822     0.00000
#> 1977-1981    7.853021     0.00000     0.00000
#> 
#> 
#> [[3]]
#> [[3]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      21.50538    17.43067    14.03509
#> 1972-1976    11.17939    10.54057     0.00000
#> 1977-1981     6.75725     0.00000     0.00000
#> 
#> [[3]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.533937    5.339367    3.755656
#> 1972-1976    2.668760    2.982732    0.000000
#> 1977-1981    2.074689    0.000000    0.000000
#> 
#> [[3]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     18.971440    12.09131    10.27943
#> 1972-1976    8.510634     7.55784     0.00000
#> 1977-1981    4.682562     0.00000     0.00000
#> 
#> [[3]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     19.423559    5.639098    3.132832
#> 1972-1976   14.212155    2.550900    0.000000
#> 1977-1981    3.546477    0.000000    0.000000
#> 
#> [[3]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      21.50538    17.43067    14.03509
#> 1972-1976    11.17939    10.54057     0.00000
#> 1977-1981     6.75725     0.00000     0.00000
#> 
#> 
#> [[4]]
#> [[4]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.316355    20.82626    9.734012
#> 1972-1976    8.783810    11.81822    0.000000
#> 1977-1981    4.930966     0.00000    0.000000
#> 
#> [[4]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.714932    4.072398    3.891403
#> 1972-1976    2.080063    2.943485    0.000000
#> 1977-1981    1.625173    0.000000    0.000000
#> 
#> [[4]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.601423   16.753861     5.84261
#> 1972-1976    6.703747    8.874732     0.00000
#> 1977-1981    3.305794    0.000000     0.00000
#> 
#> [[4]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     28.195489    8.771930    5.012531
#> 1972-1976   13.847740    3.279728    0.000000
#> 1977-1981    3.103167    0.000000    0.000000
#> 
#> [[4]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.316355    20.82626    9.734012
#> 1972-1976    8.783810    11.81822    0.000000
#> 1977-1981    4.930966     0.00000    0.000000
#> 
#> 
#> [[5]]
#> [[5]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     27.164686    22.18449    11.77136
#> 1972-1976    9.422632    12.45704     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#> [[5]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      3.076923    4.253394    3.710407
#> 1972-1976    2.472527    2.904239    0.000000
#> 1977-1981    1.521438    0.000000    0.000000
#> 
#> [[5]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     24.087763   17.931100    8.060957
#> 1972-1976    6.950105    9.552801    0.000000
#> 1977-1981    6.879467    0.000000    0.000000
#> 
#> [[5]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     18.796992    5.639098    3.446115
#> 1972-1976   12.025669    2.550900    0.000000
#> 1977-1981    6.206334    0.000000    0.000000
#> 
#> [[5]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     27.164686    22.18449    11.77136
#> 1972-1976    9.422632    12.45704     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#>