A Function to Sample the Annual Percentage Decrease with Uncertainty

A Function to Sample the Annual Percentage Decrease with Uncertainty

sample_per_decrease(df, max_year = NULL, min_year = NULL,
  samples = 100, years_to_avg = 3)

Arguments

df
A dataframe containing a year variable and with the second variable being the one to sample from.

max_year
Numeric, set as greater than the maximum year in the dataframe if interested in projection. Defaults to NULL. If specified, then an average of the percentage decrease for a set number of years is taken. This is specified with years_to_avg.

min_year
Numeric, set as smaller than the minimum year in the dataframe if interested in backwards projection. Defaults to NULL. If specified, then an average of the percentage decrease for a set number of years is taken. This is specified with years_to_avg.

samples
Numeric, the number of samples to take.

years_to_avg
Numeric, the number of years over which to average the percentage decrease when estimating the ongoing percentage decrease.

Value

A dataframe containing annual percentage decrease samples.

Examples

 df <- data.frame(year = c(2000:2010), measure = seq(10, 110, 10))

sample_per_decrease(df, samples = 2)

#>    year sample prop_decrease_sample
#> 1  2001      1          -0.86905934
#> 2  2001      2          -0.96799015
#> 3  2002      1          -0.31209773
#> 4  2002      2          -0.46914099
#> 5  2003      1          -0.28566263
#> 6  2003      2          -0.40189774
#> 7  2004      1          -0.14339555
#> 8  2004      2          -0.32471687
#> 9  2005      1          -0.22280581
#> 10 2005      2          -0.16012861
#> 11 2006      1          -0.18337064
#> 12 2006      2          -0.12078320
#> 13 2007      1          -0.17205021
#> 14 2007      2          -0.14077388
#> 15 2008      1          -0.09130997
#> 16 2008      2          -0.09451527
#> 17 2009      1          -0.15581897
#> 18 2009      2          -0.08803954
#> 19 2010      1          -0.11111544
#> 20 2010      2          -0.07641219