Data used in Sutherland et al

A dataset containing the data used by Sutherland et al to estimate the impact of ending the BCG school's scheme. The data has been aggregated for surveys. Details of its collection can be found in the original paper, in sutherland_data.R (see data-raw) and in the vignette discussing recreating Sutherland et al.'s model.

sutherland_data

Format

A list with 14 elements.

Cohort

A character string indicating the year of entry to study

Cohort.Ranges

Age ranges included in the study; 15-19, 20-24 and '25-29 years

Projected.Cohort

The number of years for which to project forward (scaled by the length of cohort)

Cohort.Pop

The population at entry in each cohort

Unvaccinated

Incidence rates (per 100,000) in those unvaccinated by cohort and age group

Vaccinated

Incidence rates (per 100,000) in those vaccinated by cohort and age group

Notifications.Prevented

Notifications prevented by 100,000 people by cohor and age groups

BCG.efficacy

Estimated efficacy of the BCG vaccine

BCG.vaccinations

Number of BCG vaccines given by 5 year period

BCG.Unvaccinated

Number of BCG unvaccinated by 5 year period

BCG.Tuberculin

Number Tuberculin positive by 5 year period

BCG.Ineglible

Number inelgible for the BCG schools scheme by 5 year period