Incidence rates used in Sutherland et al

A dataset containing the incidence rates used by Sutherland et al to estimate the impact of ending the BCG school's scheme. The data has been aggregated for surveys. Details of its collection can be found in the original paper, in sutherland_incidence_rates.R (see data-raw) and in the vignette discussing recreating Sutherland et al.'s model.

sutherland_incidence_rates

Format

A list with 14 elements.

Unvaccinated

A matrix of incidence rates for the unvaccinated by age and year of entry

Vaccinated

A matrix of incidence rates for the vaccinated by age and year of entry

Notifications.Prevented

A matrix of notifications prevented (as rates) by age and year of entry

Tuberculin

Incidence rates (per 100,000) in the Tuberculin positive population

Inelgible

Incidence rates (per 100,000) in those inelgible for BCG vaccination