Sutherland et al. Model to Estimate Impact of Ending the BCG School Scheme

Sutherland et al. Model to Estimate Impact of Ending the BCG School Scheme

sutherland_model(Data = sutherland_data,
  incidence_rates = sutherland_incidence_rates, Rates.Per = 1e+05,
  Cohort.Length = 5, Data.start = 1969,
  Annual.TB.Decrease.Yearly = TB_decrease_as_matrix(sutherland_TB_decrease),
  Percentage.Year.One = NULL, trans_params = sutherland_trans_params(),
  Sym.Lag = 2)

Arguments

Data
Data used by Sutherland et al. analysis. See sutherland_data.

incidence_rates
A list of Incidence rates (per 100,000) for the BCG vaccinated and unvaccinated population. See sutherland_incidence_rates for details.

Rates.Per
The number of people by which to produce rates, defaults to 100,000 as in Sutherland et al.

Cohort.Length
Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.

Data.start
Numeric, the year the first data point originates, defaults to 1969 as in Sutherland et al.

Annual.TB.Decrease.Yearly
Matrix of average yearly decreases in TB with each column representing an age group (15-19, 20-24, 25-29). Row names must be the year, starting from Data.start. Defaults to using sutherland_TB_decrease.

Percentage.Year.One
Numeric, the percentage of a generation that occur in the first year. Fitted via, fit_model_to_data.

trans_params
A funciton for estimating the transmission parameters (expected total secondary notifications, the size of the first generation, and the average interval between a primary case and all secondary cases). Defailts to sutherland_trans_params.

Sym.Lag
Numeric, the generation time between infection and symtoms. Must be smaller or equal to the Cohort.Length. Defaults to sutherland_gen_time.

Value

A list of tables reproducing the results presented in Sutherland et al. The final table estimates the total number of additional cases from ending the shcheme. In order to provide a complete estimate each 5 year estimate has been multiplied by the cohort length.

Examples

sutherland_model(Percentage.Year.One = 0.764)

#> $`Table 2 - Estimated risk of developing notified TB`
#>       Aged 13 in Unvaccinated Vaccinated
#>  [1,]       1969          370       1800
#>  [2,]       1974          680       2400
#>  [3,]       1979         1000       4000
#>  [4,]       1984         1700       6500
#>  [5,]       1989         2700      10000
#>  [6,]       1994         4300      17000
#>  [7,]       1999         6900      27000
#>  [8,]       2004        11000      43000
#>  [9,]       2009        18000      68000
#> 
#> $`Table 3 - Estimated of no. of TB notifications prevented`
#>       Vac at age 13 Notifications prevented in 15 years
#>  [1,]          1969                                 218
#>  [2,]          1974                                 106
#>  [3,]          1979                                  71
#>  [4,]          1984                                  44
#>  [5,]          1989                                  28
#>  [6,]          1994                                  17
#>  [7,]          1999                                  11
#>  [8,]          2004                                   7
#>  [9,]          2009                                   4
#>       Vacs to prevent one notification in 15 years
#>  [1,]                                          460
#>  [2,]                                          940
#>  [3,]                                         1400
#>  [4,]                                         2200
#>  [5,]                                         3600
#>  [6,]                                         5800
#>  [7,]                                         9300
#>  [8,]                                        15000
#>  [9,]                                        24000
#> 
#> $`Table 4 - Estimated no. of TB notifications prevented by Schools BCG scheme`
#>      15-19 years 20-24 years 25-29 years
#> 1969         449         345         171
#> 1974         207         211         123
#> 1979         175         150          87
#> 1984          96          82          48
#> 1989          47          40          24
#> 1994          31          26          15
#> 1999          19          16          10
#> 2004          12          10           6
#> 2009           8           6           4
#> 
#> $`Table 4 - Total notifcations prevented each year`
#> 1983 1988 1993 1998 2003 2008 2013 
#>  558  369  217  119   69   44   27 
#> 
#> $`Table 5 - Total notifications if the scheme continues`
#> 1983 1988 1993 1998 2003 2008 2013 
#>  463  296  166   91   52   33   22 
#> 
#> $`Table 5 - Primary additional notifications`
#>      1973 1978 1983 1988 1993 1998 2003 2008 2013
#> 1971    0  207  387  369  217  119   69   44   27
#> 1976    0    0  175  246  217  119   69   44   27
#> 1981    0    0    0   96  130  119   69   44   27
#> 1986    0    0    0    0   47   71   69   44   27
#> 1991    0    0    0    0    0   31   46   44   27
#> 1996    0    0    0    0    0    0   19   29   27
#> 2001    0    0    0    0    0    0    0   12   18
#> 2006    0    0    0    0    0    0    0    0    8
#> 2011    0    0    0    0    0    0    0    0    0
#> 
#> $`Table 5 - Secondary additional notifications`
#>      1978 1983 1988 1993 1998 2003 2008 2013
#> 1971   59  209  290  238  138   77   46   29
#> 1976    0   50  154  180  138   77   46   29
#> 1981    0    0   28   83   96   77   46   29
#> 1986    0    0    0   13   43   54   46   29
#> 1991    0    0    0    0    9   28   34   29
#> 1996    0    0    0    0    0    5   18   21
#> 2001    0    0    0    0    0    0    3   11
#> 2006    0    0    0    0    0    0    0    2
#> 2011    0    0    0    0    0    0    0    0
#> 
#> $`Table 5 - Secondary additional notifcations reduced to Sutherland for clarity`
#>      1993 1998 2003 2008 2013
#> 1986   13   43   54   46   29
#> 1991    0    9   28   34   29
#> 1996    0    0    5   18   21
#> 
#> $`Table 5 - Total secondary notifications all time`
#> 1971 1976 1981 1986 1991 1996 2001 2006 2011 
#> 1086  674  359  185  100   44   14    2    0 
#> 
#> $`Table 5 - Total Effects of ending the schools BCG scheme`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`Total additional notifications`
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971  596  659  455  257  146   90   56
#> 1976  225  400  397  257  146   90   56
#> 1981    0  124  213  215  146   90   56
#> 1986    0    0   60  114  123   90   56
#> 1991    0    0    0   40   74   78   56
#> 1996    0    0    0    0   24   47   48
#> 2001    0    0    0    0    0   15   29
#> 2006    0    0    0    0    0    0   10
#> 2011    0    0    0    0    0    0    0
#> 
#> $`Total additional notifications all time`
#>  1971  1976  1981  1986  1991  1996  2001  2006  2011 
#> 11295  7855  4220  2215  1240   595   220    50     0 
#>