Estimate Transmission chain parameters using Sutherland et al.s Approach

Estimate Transmission chain parameters using Sutherland et al.s Approach

sutherland_trans_params(Annual.TB.Decrease.Yearly = TB_decrease_as_matrix(sutherland_TB_decrease),
  Sym.Lag = sutherland_gen_time, Sutherland = TRUE, Repro.No = NULL,
  Cohort.Length = 5)

Arguments

Annual.TB.Decrease.Yearly
Matrix of average yearly decreases in TB with each column representing an age group (15-19, 20-24, 25-29). Row names must be the year, starting from Data.start. Defaults to using sutherland_TB_decrease.

Sym.Lag
Numeric, the generation time between infection and symtoms. Must be smaller or equal to the Cohort.Length. Defaults to sutherland_gen_time.

Sutherland
Logical, defaults to TRUE. Whether or not to use Sutherland et al.'s assumption for estimating the basic reproduction number.

Repro.No
If Sutherland is FALSE sets the basic reproduction number used to estimate the transmission chain parameters.

Cohort.Length
Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.

Value

A dlist containing the expected total secondary notifications (Expected.Total.Sec.Note), the size of the first generation (Size.First.Gen) and the average interval to all secondary cases (Avg.Int.All.Sec).

Examples

sutherland_trans_params()

#> $Expected.Total.Sec.Note
#>   1969   1970   1971   1972   1973   1974   1975   1976   1977   1978   1979 
#> 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 
#>   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989   1990 
#> 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 
#>   1991   1992   1993   1994   1995   1996   1997   1998   1999   2000   2001 
#> 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 
#>   2002   2003   2004   2005   2006   2007   2008   2009   2010   2011   2012 
#> 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 0.8281 
#>   2013 
#> 0.8281 
#> 
#> $Size.First.Gen
#>     1969     1970     1971     1972     1973     1974     1975     1976 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 
#>     1977     1978     1979     1980     1981     1982     1983     1984 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 
#>     1985     1986     1987     1988     1989     1990     1991     1992 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 
#>     1993     1994     1995     1996     1997     1998     1999     2000 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 
#>     2001     2002     2003     2004     2005     2006     2007     2008 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 0.452984 
#>     2009     2010     2011     2012     2013 
#> 0.452984 0.452984 0.452984 0.452984 0.452984 
#> 
#> $Avg.Int.All.Sec
#>   1969   1970   1971   1972   1973   1974   1975   1976   1977   1978   1979 
#> 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 
#>   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989   1990 
#> 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 
#>   1991   1992   1993   1994   1995   1996   1997   1998   1999   2000   2001 
#> 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 
#>   2002   2003   2004   2005   2006   2007   2008   2009   2010   2011   2012 
#> 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 3.6562 
#>   2013 
#> 3.6562 
#>