Function to sample Sutherland Incidence rates to Include Uncertainty

Function to sample Sutherland Incidence rates to Include Uncertainty

sample_incidence_rates(incidence_rates, population_data, samples = 100)

Arguments

incidence_rates
Incidence rates as formated by sutherland_incidence_rates

population_data
Population data as formated by sutherland_data

samples
Numeric, the number of samples to take. Defaults to 100.

Value

A list of lists. Each list contains matrices of sampled incidence rates for unvaccinated, vaccinated, tuberculin postive and inelgible populations.

Examples

sample_incidence_rates(sutherland_incidence_rates, sutherland_data, samples = 5)

#> [[1]]
#> [[1]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     27.391058    22.18449    11.77136
#> 1972-1976   15.172035    12.45704     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#> [[1]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      3.076923    4.253394    3.710407
#> 1972-1976    2.472527    2.904239    0.000000
#> 1977-1981    1.521438    0.000000    0.000000
#> 
#> [[1]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     24.314135   17.931100    8.060957
#> 1972-1976   12.699507    9.552801    0.000000
#> 1977-1981    6.879467    0.000000    0.000000
#> 
#> [[1]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     17.543860    6.265664    3.446115
#> 1972-1976   17.127468    5.466213    0.000000
#> 1977-1981    3.103167    0.000000    0.000000
#> 
#> [[1]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     27.391058    22.18449    11.77136
#> 1972-1976   15.172035    12.45704     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#> 
#> [[2]]
#> [[2]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     22.637238    21.50538    9.281268
#> 1972-1976   11.019689    10.22116    0.000000
#> 1977-1981    7.853021     0.00000    0.000000
#> 
#> [[2]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      1.945701    5.113122    3.936652
#> 1972-1976    2.237049    2.864992    0.000000
#> 1977-1981    1.798064    0.000000    0.000000
#> 
#> [[2]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.691537   16.392254    5.344616
#> 1972-1976    8.782640    7.356168    0.000000
#> 1977-1981    6.054957    0.000000    0.000000
#> 
#> [[2]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     21.929825    6.578947    3.446115
#> 1972-1976   10.568012    5.466213    0.000000
#> 1977-1981    6.206334    0.000000    0.000000
#> 
#> [[2]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     22.637238    21.50538    9.281268
#> 1972-1976   11.019689    10.22116    0.000000
#> 1977-1981    7.853021     0.00000    0.000000
#> 
#> 
#> [[3]]
#> [[3]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.316355    19.01528    12.67685
#> 1972-1976   10.061455    10.22116     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#> [[3]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.398190    5.158371    3.800905
#> 1972-1976    3.453689    2.982732    0.000000
#> 1977-1981    1.659751    0.000000    0.000000
#> 
#> [[3]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.918165   13.856909    8.875948
#> 1972-1976    6.607766    7.238429    0.000000
#> 1977-1981    6.741155    0.000000    0.000000
#> 
#> [[3]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.989975    6.892231    3.446115
#> 1972-1976   14.940983    2.186485    0.000000
#> 1977-1981    3.989786    0.000000    0.000000
#> 
#> [[3]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.316355    19.01528    12.67685
#> 1972-1976   10.061455    10.22116     0.00000
#> 1977-1981    8.400906     0.00000     0.00000
#> 
#> 
#> [[4]]
#> [[4]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      22.63724    24.44822    11.31862
#> 1972-1976    10.70028    10.85998     0.00000
#> 1977-1981    10.40982     0.00000     0.00000
#> 
#> [[4]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.352941    4.977376    3.393665
#> 1972-1976    2.394035    3.257457    0.000000
#> 1977-1981    1.417704    0.000000    0.000000
#> 
#> [[4]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.284297   19.470842    7.924954
#> 1972-1976    8.306243    7.602526    0.000000
#> 1977-1981    8.992114    0.000000    0.000000
#> 
#> [[4]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     16.917293    4.385965    4.385965
#> 1972-1976   18.220711    2.915314    0.000000
#> 1977-1981    6.206334    0.000000    0.000000
#> 
#> [[4]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      22.63724    24.44822    11.31862
#> 1972-1976    10.70028    10.85998     0.00000
#> 1977-1981    10.40982     0.00000     0.00000
#> 
#> 
#> [[5]]
#> [[5]]$Unvaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.995473    18.33616    11.77136
#> 1972-1976    8.304693    10.70028     0.00000
#> 1977-1981    8.035649     0.00000     0.00000
#> 
#> [[5]]$Vaccinated
#>           15-19 years 20-24 years 25-29 years
#> 1967-71      2.714932    4.570136    4.208145
#> 1972-1976    3.375196    3.375196    0.000000
#> 1977-1981    1.798064    0.000000    0.000000
#> 
#> [[5]]$Notifications.Prevented
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     21.280540   13.766027    7.563219
#> 1972-1976    4.929497    7.325081    0.000000
#> 1977-1981    6.237585    0.000000    0.000000
#> 
#> [[5]]$Tuberculin
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     20.363409    7.205514    3.446115
#> 1972-1976   19.313954    2.186485    0.000000
#> 1977-1981    4.876405    0.000000    0.000000
#> 
#> [[5]]$Inelgible
#>           15-19 years 20-24 years 25-29 years
#> 1967-71     23.995473    18.33616    11.77136
#> 1972-1976    8.304693    10.70028     0.00000
#> 1977-1981    8.035649     0.00000     0.00000
#> 
#>