Function to Sample Model Parameters

The serial interval is estimated as 1.44 (95 this study. Samples were taken by assuming a normal distribution with a standard deviation estimated from the parameters confidence intervals. The above study includes data on the percentage of cases in each year since infection in the form of a figure. This appears to roughly match the original model estimate. As no other estimate is available the point estimate from the original model is used.

sample_model_parameters(incidence_rates = sutherland_incidence_rates,
  population_data = sutherland_data, samples = 100,
  as_sutherland = FALSE)

Arguments

incidence_rates
Incidence rates as formated by sutherland_incidence_rates

population_data
Population data as formated by sutherland_data

samples
Numeric, the number of samples to run. Defaults to 100

as_sutherland
Logical, defaults to FALSE. If TRUE then the original estimates used by sutherland et al. are returned. Note these are point estimates.

Value

A tibble of parameter samples

See also

Examples

sample_model_parameters()

#> # A tibble: 100 x 4
#>    sample incidence_rates per_year_one sym_lag
#>     <int> <list>                 <dbl>   <dbl>
#>  1      1 <list [5]>             0.764    1.40
#>  2      2 <list [5]>             0.764    1.39
#>  3      3 <list [5]>             0.764    1.48
#>  4      4 <list [5]>             0.764    1.51
#>  5      5 <list [5]>             0.764    1.39
#>  6      6 <list [5]>             0.764    1.42
#>  7      7 <list [5]>             0.764    1.45
#>  8      8 <list [5]>             0.764    1.50
#>  9      9 <list [5]>             0.764    1.39
#> 10     10 <list [5]>             0.764    1.37
#> # … with 90 more rows