Function to run multiple simulations of the Sutherland et al. Estimation Model

Function to run multiple simulations of the Sutherland et al. Estimation Model

simulate_sutherland_model(list_annual_TB_decrease = list(TB_decrease_as_matrix(sutherland_TB_decrease),
  TB_decrease_as_matrix(sutherland_TB_decrease)),
  list_incidence_rates = list(sutherland_incidence_rates,
  sutherland_incidence_rates), list_percentage_year_one = rep(0.764, 2),
  list_sym_lag = list(2, 2), cores = 1, ...)

Arguments

list_annual_TB_decrease
A list of annual TB decreases with each element as specified for the Annual.TB.Decrease.Yearly arguement of sutherland_model.

list_incidence_rates
A list of incidence rates with each element as specified for the incidence_rates arguement of sutherland_model.

list_percentage_year_one
A list of percentage secondary notifications in the first year with each element as specified for the Percentage.Year.One arguement of sutherland_model.

cores
The number of cores to use, defaults to one

...
Additional parameters to pass to sutherland_model.

Value

A list of simulations for each of the outputs of sutherland_model.

Examples

simulate_sutherland_model(cores = 1)

#> $`Table 2 - Estimated risk of developing notified TB`
#> $`Table 2 - Estimated risk of developing notified TB`[[1]]
#>       Aged 13 in Unvaccinated Vaccinated
#>  [1,]       1969          370       1800
#>  [2,]       1974          680       2400
#>  [3,]       1979         1000       4000
#>  [4,]       1984         1700       6500
#>  [5,]       1989         2700      10000
#>  [6,]       1994         4300      17000
#>  [7,]       1999         6900      27000
#>  [8,]       2004        11000      43000
#>  [9,]       2009        18000      68000
#> 
#> $`Table 2 - Estimated risk of developing notified TB`[[2]]
#>       Aged 13 in Unvaccinated Vaccinated
#>  [1,]       1969          370       1800
#>  [2,]       1974          680       2400
#>  [3,]       1979         1000       4000
#>  [4,]       1984         1700       6500
#>  [5,]       1989         2700      10000
#>  [6,]       1994         4300      17000
#>  [7,]       1999         6900      27000
#>  [8,]       2004        11000      43000
#>  [9,]       2009        18000      68000
#> 
#> 
#> $`Table 3 - Estimated of no. of TB notifications prevented`
#> $`Table 3 - Estimated of no. of TB notifications prevented`[[1]]
#>       Vac at age 13 Notifications prevented in 15 years
#>  [1,]          1969                                 218
#>  [2,]          1974                                 106
#>  [3,]          1979                                  71
#>  [4,]          1984                                  44
#>  [5,]          1989                                  28
#>  [6,]          1994                                  17
#>  [7,]          1999                                  11
#>  [8,]          2004                                   7
#>  [9,]          2009                                   4
#>       Vacs to prevent one notification in 15 years
#>  [1,]                                          460
#>  [2,]                                          940
#>  [3,]                                         1400
#>  [4,]                                         2200
#>  [5,]                                         3600
#>  [6,]                                         5800
#>  [7,]                                         9300
#>  [8,]                                        15000
#>  [9,]                                        24000
#> 
#> $`Table 3 - Estimated of no. of TB notifications prevented`[[2]]
#>       Vac at age 13 Notifications prevented in 15 years
#>  [1,]          1969                                 218
#>  [2,]          1974                                 106
#>  [3,]          1979                                  71
#>  [4,]          1984                                  44
#>  [5,]          1989                                  28
#>  [6,]          1994                                  17
#>  [7,]          1999                                  11
#>  [8,]          2004                                   7
#>  [9,]          2009                                   4
#>       Vacs to prevent one notification in 15 years
#>  [1,]                                          460
#>  [2,]                                          940
#>  [3,]                                         1400
#>  [4,]                                         2200
#>  [5,]                                         3600
#>  [6,]                                         5800
#>  [7,]                                         9300
#>  [8,]                                        15000
#>  [9,]                                        24000
#> 
#> 
#> $`Table 4 - Estimated no. of TB notifications prevented by Schools BCG scheme`
#> $`Table 4 - Estimated no. of TB notifications prevented by Schools BCG scheme`[[1]]
#>      15-19 years 20-24 years 25-29 years
#> 1969         449         345         171
#> 1974         207         211         123
#> 1979         175         150          87
#> 1984          96          82          48
#> 1989          47          40          24
#> 1994          31          26          15
#> 1999          19          16          10
#> 2004          12          10           6
#> 2009           8           6           4
#> 
#> $`Table 4 - Estimated no. of TB notifications prevented by Schools BCG scheme`[[2]]
#>      15-19 years 20-24 years 25-29 years
#> 1969         449         345         171
#> 1974         207         211         123
#> 1979         175         150          87
#> 1984          96          82          48
#> 1989          47          40          24
#> 1994          31          26          15
#> 1999          19          16          10
#> 2004          12          10           6
#> 2009           8           6           4
#> 
#> 
#> $`Table 4 - Total notifcations prevented each year`
#> $`Table 4 - Total notifcations prevented each year`[[1]]
#> 1983 1988 1993 1998 2003 2008 2013 
#>  558  369  217  119   69   44   27 
#> 
#> $`Table 4 - Total notifcations prevented each year`[[2]]
#> 1983 1988 1993 1998 2003 2008 2013 
#>  558  369  217  119   69   44   27 
#> 
#> 
#> $`Table 5 - Total notifications if the scheme continues`
#> $`Table 5 - Total notifications if the scheme continues`[[1]]
#> 1983 1988 1993 1998 2003 2008 2013 
#>  463  296  166   91   52   33   22 
#> 
#> $`Table 5 - Total notifications if the scheme continues`[[2]]
#> 1983 1988 1993 1998 2003 2008 2013 
#>  463  296  166   91   52   33   22 
#> 
#> 
#> $`Table 5 - Primary additional notifications`
#> $`Table 5 - Primary additional notifications`[[1]]
#>      1973 1978 1983 1988 1993 1998 2003 2008 2013
#> 1971    0  207  387  369  217  119   69   44   27
#> 1976    0    0  175  246  217  119   69   44   27
#> 1981    0    0    0   96  130  119   69   44   27
#> 1986    0    0    0    0   47   71   69   44   27
#> 1991    0    0    0    0    0   31   46   44   27
#> 1996    0    0    0    0    0    0   19   29   27
#> 2001    0    0    0    0    0    0    0   12   18
#> 2006    0    0    0    0    0    0    0    0    8
#> 2011    0    0    0    0    0    0    0    0    0
#> 
#> $`Table 5 - Primary additional notifications`[[2]]
#>      1973 1978 1983 1988 1993 1998 2003 2008 2013
#> 1971    0  207  387  369  217  119   69   44   27
#> 1976    0    0  175  246  217  119   69   44   27
#> 1981    0    0    0   96  130  119   69   44   27
#> 1986    0    0    0    0   47   71   69   44   27
#> 1991    0    0    0    0    0   31   46   44   27
#> 1996    0    0    0    0    0    0   19   29   27
#> 2001    0    0    0    0    0    0    0   12   18
#> 2006    0    0    0    0    0    0    0    0    8
#> 2011    0    0    0    0    0    0    0    0    0
#> 
#> 
#> $`Table 5 - Secondary additional notifications`
#> $`Table 5 - Secondary additional notifications`[[1]]
#>      1978 1983 1988 1993 1998 2003 2008 2013
#> 1971   59  209  290  238  138   77   46   29
#> 1976    0   50  154  180  138   77   46   29
#> 1981    0    0   28   83   96   77   46   29
#> 1986    0    0    0   13   43   54   46   29
#> 1991    0    0    0    0    9   28   34   29
#> 1996    0    0    0    0    0    5   18   21
#> 2001    0    0    0    0    0    0    3   11
#> 2006    0    0    0    0    0    0    0    2
#> 2011    0    0    0    0    0    0    0    0
#> 
#> $`Table 5 - Secondary additional notifications`[[2]]
#>      1978 1983 1988 1993 1998 2003 2008 2013
#> 1971   59  209  290  238  138   77   46   29
#> 1976    0   50  154  180  138   77   46   29
#> 1981    0    0   28   83   96   77   46   29
#> 1986    0    0    0   13   43   54   46   29
#> 1991    0    0    0    0    9   28   34   29
#> 1996    0    0    0    0    0    5   18   21
#> 2001    0    0    0    0    0    0    3   11
#> 2006    0    0    0    0    0    0    0    2
#> 2011    0    0    0    0    0    0    0    0
#> 
#> 
#> $`Table 5 - Secondary additional notifcations reduced to Sutherland for clarity`
#> $`Table 5 - Secondary additional notifcations reduced to Sutherland for clarity`[[1]]
#>      1993 1998 2003 2008 2013
#> 1986   13   43   54   46   29
#> 1991    0    9   28   34   29
#> 1996    0    0    5   18   21
#> 
#> $`Table 5 - Secondary additional notifcations reduced to Sutherland for clarity`[[2]]
#>      1993 1998 2003 2008 2013
#> 1986   13   43   54   46   29
#> 1991    0    9   28   34   29
#> 1996    0    0    5   18   21
#> 
#> 
#> $`Table 5 - Total secondary notifications all time`
#> $`Table 5 - Total secondary notifications all time`[[1]]
#> 1971 1976 1981 1986 1991 1996 2001 2006 2011 
#> 1086  674  359  185  100   44   14    2    0 
#> 
#> $`Table 5 - Total secondary notifications all time`[[2]]
#> 1971 1976 1981 1986 1991 1996 2001 2006 2011 
#> 1086  674  359  185  100   44   14    2    0 
#> 
#> 
#> $`Table 5 - Total Effects of ending the schools BCG scheme`
#> $`Table 5 - Total Effects of ending the schools BCG scheme`[[1]]
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> $`Table 5 - Total Effects of ending the schools BCG scheme`[[2]]
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971 1059  955  621  348  198  123   78
#> 1976  688  696  563  348  198  123   78
#> 1981  463  420  379  306  198  123   78
#> 1986  463  296  226  205  175  123   78
#> 1991  463  296  166  131  126  111   78
#> 1996  463  296  166   91   76   80   70
#> 2001  463  296  166   91   52   48   51
#> 2006  463  296  166   91   52   33   32
#> 2011  463  296  166   91   52   33   22
#> 
#> 
#> $`Total additional notifications`
#> $`Total additional notifications`[[1]]
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971  596  659  455  257  146   90   56
#> 1976  225  400  397  257  146   90   56
#> 1981    0  124  213  215  146   90   56
#> 1986    0    0   60  114  123   90   56
#> 1991    0    0    0   40   74   78   56
#> 1996    0    0    0    0   24   47   48
#> 2001    0    0    0    0    0   15   29
#> 2006    0    0    0    0    0    0   10
#> 2011    0    0    0    0    0    0    0
#> 
#> $`Total additional notifications`[[2]]
#>      1983 1988 1993 1998 2003 2008 2013
#> 1971  596  659  455  257  146   90   56
#> 1976  225  400  397  257  146   90   56
#> 1981    0  124  213  215  146   90   56
#> 1986    0    0   60  114  123   90   56
#> 1991    0    0    0   40   74   78   56
#> 1996    0    0    0    0   24   47   48
#> 2001    0    0    0    0    0   15   29
#> 2006    0    0    0    0    0    0   10
#> 2011    0    0    0    0    0    0    0
#> 
#> 
#> $`Total additional notifications all time`
#> $`Total additional notifications all time`[[1]]
#>  1971  1976  1981  1986  1991  1996  2001  2006  2011 
#> 11295  7855  4220  2215  1240   595   220    50     0 
#> 
#> $`Total additional notifications all time`[[2]]
#>  1971  1976  1981  1986  1991  1996  2001  2006  2011 
#> 11295  7855  4220  2215  1240   595   220    50     0 
#> 
#>